moved visited schools view to a seperate component
