+++ src/app/children/child-details/child-details.component.html
@@ -236,21 +236,10 @@
 </mat-expansion-panel>
 
 <mat-expansion-panel expanded="false">
-  <mat-expansion-panel-header class="section-header" (click)="loadSchoolEntries()">
+  <mat-expansion-panel-header class="section-header">
     Schools
   </mat-expansion-panel-header>
-
-  <button mat-stroked-button color="accent" (click)="addSchoolClick()">
-    <mat-icon class="button-icon" aria-label="add child" fontIcon="fa-plus-circle"></mat-icon>Add School
-  </button>
-
-  <mat-list *ngFor="let s of viewableSchools" (click)="schoolClicked(s)" >
-    <mat-list-item >
-      <label><strong>SchoolName:</strong> {{s.school.name}} </label>
-      <label><strong> From:</strong> {{s.childSchoolRelation.start ? s.childSchoolRelation.start : 'click to enter'}} </label>
-      <label><strong> To:</strong> {{s.childSchoolRelation.end ? s.childSchoolRelation.end : 'click to enter'}} </label>
-    </mat-list-item>
-  </mat-list>
+  <app-view-schools-component [child]="child"></app-view-schools-component>
 </mat-expansion-panel>
