+++ src/app/children/view-schools-component/edit-school-dialog/edit-school-dialog.component.ts
@@ -1,11 +1,11 @@
 import {Component, Inject} from '@angular/core';
-import {EntityMapperService} from "../../entity/entity-mapper.service";
-import {ChildSchoolRelation} from "../childSchoolRelation";
-import { School } from "../../schools/school";
+import {EntityMapperService} from "../../../entity/entity-mapper.service";
+import {ChildSchoolRelation} from "../../childSchoolRelation";
+import { School } from "../../../schools/school";
 
 import uniqid from 'uniqid';
 import {MAT_DIALOG_DATA, MatDialogRef} from "@angular/material";
-import {Child} from "../child";
+import {Child} from "../../child";
 
 @Component({
   selector: 'app-add-school-dialog',
